requirements.txt
openai-whisper
transformers
pyannote.audio
yt_dlp